+++ lizard_progress/util/directories.py
@@ -38,6 +38,10 @@ def activity_dir(activity):
         project_dir(activity.project), str(activity.id)))
 
 
+def upload_dir(activity):
+    return mk(os.path.join(activity_dir(activity), 'uploads'))
+
+
 def make_uploaded_file_path(root, activity, filename):
     """Gives the path to some uploaded file, which depends on the
     project it is for, the contractor that uploaded it and the
@@ -48,11 +52,21 @@ def make_uploaded_file_path(root, activity, filename):
     Nginx for X-Sendfile (uses /protected/ as the root).
 
     External URLs should use a reverse() call to the
-    lizard_progress_filedownload view instead of this function."""
+    lizard_progress_filedownload view instead of this function
 
-    return os.path.join(
-        activity_dir(activity), 'uploads',
-        os.path.basename(filename)).replace(BASE_DIR, root)
+    NOTE: Due to a bug, many files are saved directly in the activity dir,
+    instead of in the upload dir. If the wrong one exists and the new
+    one doesn't, return the old one."""
+
+    old_path = os.path.join(activity_dir(activity), filename)
+    new_path = os.path.join(upload_dir(activity), filename)
+
+    if not os.path.exists(new_path) and os.path.exists(old_path):
+        path = old_path
+    else:
+        path = new_path
+
+    return path.replace(BASE_DIR, root)
 
 
 def results_dir(activity):