+++ CHANGES.rst
@@ -24,6 +24,10 @@ New features for the Almere / HDSR project:
 
 - Add expected attachments to popup, if present
 
+- Fix bug where uploaded files were saved to the wrong directory.
+
+- Improve date formatting
+
 
 2.4.7 (2015-05-08)
 ------------------